# auth-server/openapi.yaml
openapi: 3.1.0
info:
  title: MaruPhoto Auth Server API
  version: 0.1.0
  description: API contract for app auth, device registration, and photo backup.
servers:
  - url: https://api.example.com
tags:
  - name: Auth
  - name: Devices
  - name: Uploads
  - name: Assets
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    FirebaseAuthRequest:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    RegisterDeviceRequest:
      type: object
      required: [deviceId, platform, appVersion]
      properties:
        deviceId:
          type: string
        platform:
          type: string
          enum: [ios, android]
        appVersion:
          type: string
        pushToken:
          type: string
    Device:
      type: object
      required: [id, deviceId, platform, appVersion]
      properties:
        id:
          type: string
        deviceId:
          type: string
        platform:
          type: string
        appVersion:
          type: string
    InitiateUploadRequest:
      type: object
      required:
        - deviceId
        - fileName
        - fileSize
        - mimeType
        - sha256
      properties:
        deviceId:
          type: string
        localId:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        mimeType:
          type: string
        sha256:
          type: string
        capturedAt:
          type: string
          format: date-time
    UploadSession:
      type: object
      required:
        - uploadId
        - chunkSize
        - uploadUrl
        - expiresAt
      properties:
        uploadId:
          type: string
        chunkSize:
          type: integer
        uploadUrl:
          type: string
          description: Your backend upload endpoint, not a Microsoft URL.
        expiresAt:
          type: string
          format: date-time
    CompleteUploadRequest:
      type: object
      required: [uploadId]
      properties:
        uploadId:
          type: string
    Asset:
      type: object
      required: [id, fileName, mimeType, fileSize, status]
      properties:
        id:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        capturedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [processing, ready, failed]
paths:
  /v1/auth/firebase:
    post:
      tags: [Auth]
      summary: Exchange a Firebase ID token for an app session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FirebaseAuthRequest'
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
  /v1/devices:
    post:
      tags: [Devices]
      summary: Register or refresh a mobile device
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '200':
          description: Device record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
  /v1/uploads/initiate:
    post:
      tags: [Uploads]
      summary: Begin a resumable upload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiateUploadRequest'
      responses:
        '200':
          description: Upload session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSession'
  /v1/uploads/{uploadId}/chunk:
    put:
      tags: [Uploads]
      summary: Upload one chunk to backend for immediate forwarding to Graph
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: uploadId
          required: true
          schema:
            type: string
        - in: header
          name: Content-Range
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '202':
          description: Chunk accepted
  /v1/uploads/complete:
    post:
      tags: [Uploads]
      summary: Finalize upload and create asset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteUploadRequest'
      responses:
        '200':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
  /v1/assets:
    get:
      tags: [Assets]
      summary: List gallery assets for the signed-in user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asset'